requirements.txt
google-cloud-aiplatform>=1.36.0
vertexai>=0.0.1 
python-dotenv
pandas
scikit-learn